requirements.txt
asgiref==3.6.0
black==23.9.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
Django==4.2
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
fonttools==4.46.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.26.2
openapi-codec==1.3.2
packaging==23.0
pathspec==0.11.2
Pillow==9.4.0
platformdirs==3.10.0
PyJWT==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.2
uritemplate==4.1.1
urllib3==1.26.15
